Add HoundCI configuration file
Why?
By default, HoundCI uses it own defined rules to check for style
violations, I want it to use the rubocop and scss-lint settings instead.

How?
Add .hound.yml file to the project root.
Edit the file, setting the it to use the rubocop.yml and scss-lint.yml
file as rule guide.