Design the landing page
Why?
GIVEN that I'm a site visitor(signed up or not)
WHEN I go to the site URL
I SHOULD see the landing page

How?
Generate the Pages controller with the index action
Design the landing page in the index view
Style the page in the pages.scss file
Set the root route to point to pages#index

Finishing story "Design the landing page"
https://www.pivotaltracker.com/story/show/109427120
[#109427120]