Add Flight Search Form to the Landing Page
Why?
GIVEN that I'm a traveler
WHEN I visit the home page(landing page)
I SHOULD see the form to search for flight

How?
Creat a partial in the pages directory
Create the search form in the partial
Render the partial in the provided div in the home
Write some styling in the pages.scss to beautify the form

Completing "Add Flight Search Form to the Landing Page"
https://www.pivotaltracker.com/story/show/109427250
[finished #109427250]