Design 'All Flights' page
Why?
If a visitor want to see all the flights between different places, he/she
should be able to click on a link to see that.

How?
Allow the index route of the flights controller
Create the index.html.erb among the flight views
Load the flights in the flight controller index action
and store it in a class variable
In the index view, load the flight partial to display the loaded flights

Finishing feature "Design All Flights page"
https://www.pivotaltracker.com/story/show/110014128
[finished #110014128]