Implement Mail Send
Why?
When a traveler completes the booking process, he should be notified via
of the status of the process.

How?
generate action mailer object,
write the controller and the views for the
write the mail views
add letter_opener to the gemfile to enable opening mail with it
bundle

Completing "Implement Mail Send"
https://www.pivotaltracker.com/story/show/110016366
[finished #110016366]