Implement Manage Bookings
Why?
If a traveler has booked flight, he or she should be able to put the
booking the booking number that given to him and and search for the
booking. The passenger should be able to modify the informations in the
booking and save it.

How?
Create the view and the form to search for the booking
add the search action in the booking controller
add the route for the the form and the search

finishing feature "Manage booking" feature
https://www.pivotaltracker.com/story/show/110029954
[finished #110029954]